KTB-21 companion object classes by packages
