+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/CallbackQuery.kt
@@ -1,0 +1,12 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CallbackQuery(
+    @SerialName("data")
+    val data: String,
+    @SerialName("message")
+    val message: Message? = null,
+)
