+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/GetUpdates.kt
@@ -1,0 +1,10 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GetUpdates(
+    @SerialName("update_id")
+    val updateId: Long,
+)
