+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/InlineKeyboard.kt
@@ -1,0 +1,12 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class InlineKeyboard(
+    @SerialName("callback_data")
+    val callbackData: String,
+    @SerialName("text")
+    val text: String,
+)
