+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/Message.kt
@@ -1,0 +1,12 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Message(
+    @SerialName("text")
+    val text: String,
+    @SerialName("chat")
+    val chat: Chat,
+)
