+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/ReplyMarkup.kt
@@ -1,0 +1,10 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ReplyMarkup(
+    @SerialName("inline_keyboard")
+    val inlineKeyboard: List<List<InlineKeyboard>>,
+)
