+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/Response.kt
@@ -1,0 +1,10 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Response(
+    @SerialName("result")
+    val result: List<Update>,
+)
