+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/SendMessageRequest.kt
@@ -1,0 +1,14 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SendMessageRequest(
+    @SerialName("chat_id")
+    val chatId: Long?,
+    @SerialName("text")
+    val text: String,
+    @SerialName("reply_markup")
+    val replyMarkup: ReplyMarkup? = null,
+)
