+++ src/main/kotlin/ru/sokolova/englishtelegrambot/telegram/api/entities/Update.kt
@@ -1,0 +1,14 @@
+package ru.sokolova.englishtelegrambot.telegram.api.entities
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Update(
+    @SerialName("update_id")
+    val updateId: Long,
+    @SerialName("message")
+    val message: Message? = null,
+    @SerialName("callback_query")
+    val callbackQuery: CallbackQuery? = null,
+)
