+++ src/main/kotlin/ru/sokolova/englishtelegrambot/trainer/modal/Question.kt
@@ -1,0 +1,6 @@
+package ru.sokolova.englishtelegrambot.trainer.modal
+
+data class Question(
+    val variants: List<Word>,
+    val correctAnswer: Word,
+)
