+++ src/main/kotlin/ru/sokolova/englishtelegrambot/trainer/modal/Statistics.kt
@@ -1,0 +1,7 @@
+package ru.sokolova.englishtelegrambot.trainer.modal
+
+data class Statistics(
+    val totalCount: Int,
+    var learnedCount: Int,
+    val percent: Int,
+)
