+++ src/main/kotlin/ru/sokolova/englishtelegrambot/trainer/modal/Word.kt
@@ -1,0 +1,10 @@
+package ru.sokolova.englishtelegrambot.trainer.modal
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Word(
+    val original: String,
+    val translate: String,
+    var correctAnswersCount: Int = 0,
+)
